abstract classification
